Minimal MDNS - start a generic server implementation (#3855)
* Minimal MDNS updates: general clarity refactor, fixed some missed review comments, create a server that can work with all interfaces on linux/mac

* Add Put8 that my previous squash missed into bufbound

* Separate header ref from const version, to have it a bit more const-correct. Still have a const-cast in the writable version, however that should be more readable

* Add explanation comment

* Fix doccomment for Query::Append

* Fix clang compilation errors (signdness on cast)

* Remove commented out code

* Fix error in writing resource records (caught by future tests)

* Add writing of A/AAAA resource records and start writing some tests

* Add writing tests for IPv6 records

* Code review updates

* Remove unneeded driver file

* Add some unit tests for resource records

* Validate that write order works

* Test small buffer output

* Removed some unnecessary empty comments

* Add unit test for adding additional records

* Assert header untouched on failure to write the record

* Removed extra slash

* Start with resource record parsing

* Tests for A and AAAA records

* Ptr parsing: simple and complex

* Txt parsing test

* Restyler fixes

* Fix bad merge for bufbound constructor

* Fix bad merge for bufbound constructor

* Update bufbound interface some more

* Adjust for new API for BufBound and AutoFree/PacketBufferHandle

* Fix std::string casts in tests to make esp32 compiler happy

+++ src/lib/mdns/minimal/QName.h
@@ -49,7 +49,7 @@ public:
     SerializedQNameIterator & operator=(const SerializedQNameIterator &) = default;
 
     SerializedQNameIterator(const BytesRange validData, const uint8_t * position) :
-        mValidData(validData), mLookBehindMax(position - validData.Start()), mCurrentPosition(position)
+        mValidData(validData), mLookBehindMax(static_cast<size_t>(position - validData.Start())), mCurrentPosition(position)
     {}
 
     /// Advances to the next element in the sequence
@@ -76,7 +76,7 @@ private:
     static constexpr uint8_t kPtrMask     = 0xC0;
 
     BytesRange mValidData;
-    ptrdiff_t mLookBehindMax; // avoid loops by limiting lookbehind
+    size_t mLookBehindMax; // avoid loops by limiting lookbehind
     const uint8_t * mCurrentPosition;
     bool mIsValid = true;
